requirements.txt

Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
email-validator==2.1.1
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
reportlab==4.4.3
openpyxl==3.1.5
requests==2.32.3
